refactor: files structure was changed according to documentation recommendations
