+++ src/app/layout.tsx
@@ -6,7 +6,7 @@ import '@/styles/global.css';
 import { UserProvider } from '@/contexts/user-context';
 import { LocalizationProvider } from '@/components/core/localization-provider';
 import { ThemeProvider } from '@/components/core/theme-provider/theme-provider';
-import { ReduxProvider } from '@/redux/provider';
+import { ReduxProvider } from '@/lib/redux/provider';
 
 export const viewport = { width: 'device-width', initialScale: 1 } satisfies Viewport;
